requirements.txt
gphoto2==2.5.0
Pillow==11.0.0
